create personal record

+++ address_book_ui/src/app/interfaces/payloads/CreateRecordBody.ts
@@ -2,7 +2,7 @@ import { AddressCreateBody } from "./AddressCreateBody"
 import { ContactDetailsCreateBody } from "./ContactDetailsCreateBody"
 
 export interface CreateRecordBody {
-    userId: BigInt,
+    userId: BigInteger,
     isPersonal: boolean,
     firstName: string,
     lastName: string,
